refactor: move to summarize module
